Introduce `AbstractFullUseCaseTest` to bring `FullUseCaseIT` and `ExternalUseCaseIT` to a common parent layer
